I changed most of the errors -still a few problems
According to the JavaScript error console I have a problem on line 171
with an unexpected token <. I am not sure what I need to do to fix this.
As far as the unfinished variable on line 160 It was a while ago that I
wrote that and I wasn't sure where I was going with that. Perhaps I
changed my mind, not sure if it was crucial or not but I took it out so
there is no error.